requirements.txt
python-telegram-bot[job-queue]==20.3
nest-asyncio==1.5.8
python-dotenv
aiofiles==23.2.1